Fix descrepency between benchmarks.
Julia helpfully removes leading whitespace sometimes from multiline quotes. It seems to do it when you don't put the """ on a new line. What a strange feature. The extra space in the python string makes it report a higher KB/s which is cheating but if Python-parsimonious cheats then I must too in order to remain comparable.